move the utility class to a separate package
