+++ dev/com.ibm.ws.grpc_fat/test-applications/StoreProducerApp.war/src/com/ibm/testapp/g3store/restProducer/client/ProducerEndpointFATServlet.java
@@ -10,6 +10,8 @@
  *******************************************************************************/
 package com.ibm.testapp.g3store.restProducer.client;
 
+import static com.ibm.ws.fat.grpc.monitoring.GrpcMetricsTestUtils.checkMetric;
+import static com.ibm.ws.fat.grpc.monitoring.GrpcMetricsTestUtils.getMetric;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
@@ -436,7 +438,7 @@ public class ProducerEndpointFATServlet extends FATServlet {
 
         // retrieve the metrics
         int httpPort = Integer.parseInt(getSysProp("bvt.prop.HTTP_default"));
-        String metricValue = GrpcMetricsTestUtils.getMetric("localhost", httpPort, "/metrics/vendor/grpc.server.receivedMessages.total");
+        String metricValue = getMetric("localhost", httpPort, "/metrics/vendor/grpc.server.receivedMessages.total");
         if (metricValue == null || Integer.parseInt(metricValue) < 200) {
             fail(String.format("Incorrect metric value [%s]. Expected [%s], got [%s]", "grpc.server.receivedMessages.total", ">=200", metricValue));
         }
@@ -501,8 +503,8 @@ public class ProducerEndpointFATServlet extends FATServlet {
 
         // retrieve the metrics
         int httpPort = Integer.parseInt(getSysProp("bvt.prop.HTTP_secondary"));
-        GrpcMetricsTestUtils.checkMetric("/metrics/vendor/grpc.client.sentMessages.total", "0", "localhost", httpPort);
-        String metricValue = GrpcMetricsTestUtils.getMetric("localhost", httpPort, "/metrics/vendor/grpc.client.receivedMessages.total");
+        checkMetric("/metrics/vendor/grpc.client.sentMessages.total", "0", "localhost", httpPort);
+        String metricValue = getMetric("localhost", httpPort, "/metrics/vendor/grpc.client.receivedMessages.total");
         if (metricValue == null || Integer.parseInt(metricValue) < 200) {
             fail(String.format("Incorrect metric value [%s]. Expected [%s], got [%s]", "grpc.client.receivedMessages.total", ">=200", metricValue));
         }
